Use stack, stack is good (for overflowing)
We create per-PE chunks of size of EL3_STACK_SIZE and then use chunk as follows:
    CHUNK#1      CHUNK#2      CHUNK#3      CHUNK#4     (CPUID 0,1,2,3)
o------------o------------o------------o------------o
->context    ->context    ->context    ->context
  live stack<- live stack<- live stack<- live stack<-

Section context frame size is EL3_STORAGE_FRAME_SIZE,
the rest of EL3_STACK_SIZE is live stack.

Live stack means we use it for preserving caller registers in callee,
context is just a storage for preserving GPRs and SysRegs on SMC call.

One not ideal possibly thing is that context and live stack can collide;
Another is that EL3 and S-EL2 use the same stack.